feature/#280: rename and move component
